+++ plugins/hooks/tests/index.js
@@ -1,4 +1,5 @@
 require('./crud.js');
 require('./authz.js');
+require('./internal_event_scope.js');
 require('./email.js');
 require('./ssrf.js');